Created distributed accelerator for all ddp based implementations to inherit from. Reduce code duplication for easier accelerator integration
